pyproject.toml
[project]
name = "cv-creator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
